+++ lib/src/dart/int_extension.dart
@@ -1,54 +1,31 @@
+import 'extensions.dart';
+
 ///Int extension to extend List functionality
 extension IntExtension on int {
   /// Get Month name
-  String getMonthName([bool isShort = false]) {
-    switch (this) {
-      case 1:
-        return isShort ? 'Jan' : 'January';
-      case 2:
-        return isShort ? 'Feb' : 'February';
-      case 3:
-        return isShort ? 'Mar' : 'March';
-      case 4:
-        return isShort ? 'Apr' : 'April';
-      case 5:
-        return isShort ? 'May' : 'May';
-      case 6:
-        return isShort ? 'Jun' : 'June';
-      case 7:
-        return isShort ? 'Jul' : 'July';
-      case 8:
-        return isShort ? 'Aug' : 'August';
-      case 9:
-        return isShort ? 'Sept' : 'September';
-      case 10:
-        return isShort ? 'Oct' : 'October';
-      case 11:
-        return isShort ? 'Nov' : 'November';
-      case 12:
-        return isShort ? 'Dec' : 'December';
-    }
-    return 'Invalid day of month';
-  }
+  String getMonthName([bool isShort = false]) => match(<int, String>{
+        1: isShort ? 'Jan' : 'January',
+        2: isShort ? 'Feb' : 'February',
+        3: isShort ? 'Mar' : 'March',
+        4: isShort ? 'Apr' : 'April',
+        5: isShort ? 'May' : 'May',
+        6: isShort ? 'Jun' : 'June',
+        7: isShort ? 'Jul' : 'July',
+        8: isShort ? 'Aug' : 'August',
+        9: isShort ? 'Sept' : 'September',
+        10: isShort ? 'Oct' : 'October',
+        11: isShort ? 'Nov' : 'November',
+        12: isShort ? 'Dec' : 'December',
+      }, 'Invalid day of month');
 
   /// Get Week name
-  String getWeekName([bool isShort = false]) {
-    switch (this) {
-      case 1:
-        return isShort ? 'Mon' : 'Monday';
-      case 2:
-        return isShort ? 'tue' : 'Tuesday';
-      case 3:
-        return isShort ? 'Wed' : 'Wednesday';
-      case 4:
-        return isShort ? 'thu' : 'Thursday';
-      case 5:
-        return isShort ? 'Fri' : 'Friday';
-      case 6:
-        return isShort ? 'Sat' : 'Saturday';
-      case 7:
-        return isShort ? 'Sun' : 'Sunday';
-    }
-    return 'Invalid day of week';
-  }
+  String getWeekName([bool isShort = false]) => match(<int, String>{
+        1: isShort ? 'Mon' : 'Monday',
+        2: isShort ? 'tue' : 'Tuesday',
+        3: isShort ? 'Wed' : 'Wednesday',
+        4: isShort ? 'thu' : 'Thursday',
+        5: isShort ? 'Fri' : 'Friday',
+        6: isShort ? 'Sat' : 'Saturday',
+        7: isShort ? 'Sun' : 'Sunday',
+      }, 'Invalid day of week');
 }